requirements.txt
# Core dependencies
fastapi==0.115.12
uvicorn[standard]==0.34.0
pydantic==2.11.3
python-multipart==0.0.20

# LlamaIndex core and integrations
llama-index-core==0.12.2
llama-index==0.12.2
llama-index-llms-groq==0.3.2
llama-index-embeddings-huggingface==0.5.1
llama-index-vector-stores-chroma==0.3.1

# Vector database
chromadb==0.6.3

# Embedding model dependencies
sentence-transformers==4.0.2
torch==2.6.0
torchvision==0.21.0
torchaudio==2.6.0
transformers==4.50.3
huggingface-hub==0.30.2

# HTTP and utilities
requests==2.32.3
python-dotenv==1.1.0
getpass4==1.0.1
typing-extensions==4.13.2

# File handling
aiofiles==24.1.0
python-magic==0.4.27python-dotenv==1.2.2
